# api/docs/openapi.yaml
openapi: 3.0.3
paths:
...
  /organizations/{organization_uuid}/audit-logs:
    get:
      operationId: AuditLogsController_findAll
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
        - name: page
          required: false
          in: query
          schema:
            type: string
        - name: limit
          required: false
          in: query
          schema:
            type: string
        - name: action
          required: false
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - AuditLogs
      security:
        - bearer: []
...
  /organizations/{organization_uuid}/integrations/{integration_uuid}:
    get:
      operationId: IntegrationsController_findOne
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
        - name: integration_uuid
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - Integrations
      security:
        - bearer: []
    patch:
      operationId: IntegrationsController_update
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
        - name: integration_uuid
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateIntegrationDto"
      responses:
        "200":
          description: ""
      tags:
        - Integrations
      security:
        - bearer: []
    delete:
      operationId: IntegrationsController_delete
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
        - name: integration_uuid
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - Integrations
      security:
        - bearer: []
  /organizations/{organization_uuid}/integrations/{integration_uuid}/test:
    post:
      operationId: IntegrationsController_testConnection
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
        - name: integration_uuid
          required: true
          in: path
          schema:
            type: string
      responses:
        "201":
          description: ""
      tags:
        - Integrations
      security:
        - bearer: []
...
  /organizations/{organization_uuid}/openapi-integrations/parse:
    post:
      operationId: OpenApiIntegrationsController_parsePreview
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParseOpenApiSpecDto"
      responses:
        "201":
          description: ""
      tags:
        - OpenApiIntegrations
      security:
        - bearer: []
  /organizations/{organization_uuid}/openapi-integrations:
    post:
      operationId: OpenApiIntegrationsController_create
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOpenApiIntegrationDto"
      responses:
        "201":
          description: ""
      tags:
        - OpenApiIntegrations
      security:
        - bearer: []
  /organizations/{organization_uuid}/openapi-integrations/{integration_uuid}:
    get:
      operationId: OpenApiIntegrationsController_getDetails
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
        - name: integration_uuid
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - OpenApiIntegrations
      security:
        - bearer: []
  /organizations/{organization_uuid}/openapi-integrations/{integration_uuid}/regenerate:
    post:
      operationId: OpenApiIntegrationsController_regenerateTools
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
        - name: integration_uuid
          required: true
          in: path
          schema:
            type: string
      responses:
        "201":
          description: ""
      tags:
        - OpenApiIntegrations
      security:
        - bearer: []
  /organizations/{organization_uuid}/openapi-integrations/{integration_uuid}/test:
    post:
      operationId: OpenApiIntegrationsController_testConnection
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
        - name: integration_uuid
          required: true
          in: path
          schema:
            type: string
      responses:
        "201":
          description: ""
      tags:
        - OpenApiIntegrations
      security:
        - bearer: []
...
info:
  title: Cortex API
  description: >-
    REST API for Cortex — organizations, conversations, agents, integrations (Composio, MCP, OpenAPI, databases),
    documents, roles, and members. Authenticate via JWT Bearer token from POST /auth/email/login.
  version: 1.0.0
  contact: {}
...
servers:
  - url: http://localhost:3001
    description: Local
  - url: "{{baseUrl}}"
    description: Environment variable
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT from POST /auth/email/login
  schemas:
...
    UpdateIntegrationDto:
      type: object
      properties: {}
...
    ParseOpenApiSpecDto:
      type: object
      properties: {}
    CreateOpenApiIntegrationDto:
      type: object
      properties: {}
...
security:
  - bearer: []